serialization test should not depend on time zone
Since the assertion tests for ISO8601 formatting and ISO8601 specifies a
time zone it's necessary to ensure that the date object is in the same
timezone the serialized output is expected be in. If the time zones are
not forced to align then the test case becomes dependent on where in the
world it is being run and any potential time shifts (for example, the
observation of daylight savings time throughout most of North America).